fuzzy search

+++ models/candidate.go
@@ -1,12 +1,12 @@
 package models
 
 type Candidate struct {
-	Filename string `json:"filename"`
-	FrameID  string `json:"frame_id"`
-	Score    int    `json:"score"`
+	Filename string  `json:"filename"`
+	FrameID  string  `json:"frame_id"`
+	Score    float32 `json:"score"`
 }
 
-func NewCandidate(frame Frame, score int) *Candidate {
+func NewCandidate(frame Frame, score float32) *Candidate {
 	return &Candidate{
 		Filename: frame.GetFilename(),
 		FrameID:  frame.GetFrameID(),
